requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytest
requests
coverage
jinja2
aiohttp
pyyaml
boto3